Add tests and hide button on media edit page
