+++ src/ThunderTestTrait.php
@@ -70,4 +70,17 @@ trait ThunderTestTrait {
     }
   }
 
+  /**
+   * LogIn with defined role assigned to user.
+   *
+   * @param string $role
+   *   Role name that will be assigned to user.
+   */
+  protected function logWithRole($role) {
+    $editor = $this->drupalCreateUser();
+    $editor->addRole($role);
+    $editor->save();
+    $this->drupalLogin($editor);
+  }
+
 }
